💄Styles: Category関連のコンポーネントにstateを定義
