--- src/features/home/HomeContents.tsx
@@ -1,15 +1,0 @@
-import React from 'react';
-
-import PopularCategories from './PopularCategories';
-import TrendArticles from './TrendArticles';
-
-const HomeContents = () => {
-  return (
-    <>
-      <PopularCategories />
-      <TrendArticles />
-    </>
-  );
-};
-
-export default HomeContents;
